requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
billiard==4.2.0
blinker==1.7.0
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
flask==3.0.2
frozenlist==1.4.1
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
jsonpickle==3.0.3
kombu==5.3.5
MarkupSafe==2.1.5
multidict==6.0.5
prompt-toolkit==3.0.43
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.0.1
requests==2.31.0
six==1.16.0
slack-bolt==1.18.1
slack-sdk==3.27.0
typing-extensions==4.10.0
tzdata==2024.1
urllib3==2.2.0
vine==5.1.0
wcwidth==0.2.13
werkzeug==3.0.1
yarl==1.9.4
zipp==3.17.0
